refactor: move timeline calls to timeline repo
