+++ mobile/lib/repositories/user.repository.dart
@@ -70,26 +70,4 @@ class UserRepository extends DatabaseRepository implements IUserRepository {
       await db.users.clear();
     });
   }
-
-  @override
-  Future<List<int>> getTimelineUserIds(int id) {
-    return db.users
-        .filter()
-        .inTimelineEqualTo(true)
-        .or()
-        .isarIdEqualTo(id)
-        .isarIdProperty()
-        .findAll();
-  }
-
-  @override
-  Stream<List<int>> watchTimelineUsers(int id) {
-    return db.users
-        .filter()
-        .inTimelineEqualTo(true)
-        .or()
-        .isarIdEqualTo(id)
-        .isarIdProperty()
-        .watch();
-  }
 }
