+++ mobile/lib/services/background.service.dart
@@ -410,7 +410,6 @@ class BackgroundService {
       partnerApiRepository,
       userApiRepository,
       userRepository,
-      syncSerive,
     );
     AlbumService albumService = AlbumService(
       userService,
